requirements.txt
vanna[sqlite,gemini]>=2.0.0
fastapi
uvicorn
plotly
pandas
python-dotenv
faker
google-genai